fix: restructured different layouts
