+++ src/auth/AuthProvider.tsx
@@ -1,4 +1,4 @@
-import { AuthContext } from "@/context/auth/auth.module";
+import { AuthContext } from "@/auth/auth.module";
 import { TAuth, TLogin } from "@/types";
 import supabase from "@/utils/supabase";
 import { Session } from "@supabase/supabase-js";
@@ -9,25 +9,25 @@ type Props = {
 	children: ReactNode
 }
 
+
+
 const AuthProvider = ({ children }: Props) => {
 
 	const [session, setSession] = useState<Session | null>(null)
 
 	useEffect(() => {
-		const {data: { subscription }} = supabase.auth.onAuthStateChange(
+		const { data: { subscription } } = supabase.auth.onAuthStateChange(
 			(event, session) => {
-			  if (event === 'SIGNED_OUT') {
-				setSession(null)
-			  } else if(event === "TOKEN_REFRESHED"){
-				setSession(null)
-			  } else if (session) {
-				setSession(session)
-			  }
+				if (event === 'SIGNED_OUT') {
+					setSession(null)
+				} else if (session) {
+					setSession(session)
+				}
 			})
-		  return () => {
+		return () => {
 			subscription.unsubscribe()
-		  }
-	  }, []);
+		}
+	}, [])
 
 	// Checking if the user is admin
 	useEffect(() => {
@@ -36,19 +36,18 @@ const AuthProvider = ({ children }: Props) => {
 			try {
 				const { data, error } = await supabase.from("users_details").select("is_admin").eq("auth_id", session?.user.id).single()
 
-				if(error) throw error
+				if (error) throw error
 
-				localStorage.setItem("is_admin", data.is_admin)
 			} catch (error) {
 				console.error(error)
 
 			}
-		}  
-		
+		}
+
 		fetchUser()
 
 	}, [session])
-	
+
 
 	const login = async (payload: TLogin) => {
 		const { email, password } = payload
@@ -66,24 +65,22 @@ const AuthProvider = ({ children }: Props) => {
 
 		if (redirectTo) window.location.href = redirectTo
 
-		localStorage.removeItem("is_admin")
-		setSession(null)
 	}
 
 	useEffect(() => {
-		const {data: { subscription }} = supabase.auth.onAuthStateChange(
+		const { data: { subscription } } = supabase.auth.onAuthStateChange(
 			(event, session) => {
-			  if (event === 'SIGNED_OUT') {
-				setSession(null)
-			  } else if (session) {
-				setSession(session)
-			  }
+				if (event === 'SIGNED_OUT') {
+					setSession(null)
+				} else if (session) {
+					setSession(session)
+				}
 			})
-		  return () => {
+		return () => {
 			subscription.unsubscribe()
-		  }
-	  }, []);
-	
+		}
+	}, []);
+
 
 	const value: TAuth = {
 		login,
