+++ src/layout/AdminLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router"
+
+const AdminLayout = () => {
+
+	
+
+
+	return <Outlet/>
+}
+
+export default AdminLayout
