--- src/layout/PrivateLayout.tsx
@@ -1,11 +1,0 @@
-import { useAuth } from "@/context/auth/auth.module"
-import { Navigate, Outlet } from "react-router"
-
-const PrivateLayout = () => {
-
-	const { session } = useAuth()
-
-	return session ? <Outlet /> : <Navigate to={"/"}/>
-}
-
-export default PrivateLayout
