+++ src/pages/admin/AdminHome.tsx
@@ -1,0 +1,8 @@
+
+const AdminHome = () => {
+  return (
+	<div>Hello</div>
+  )
+}
+
+export default AdminHome
