+++ src/pages/index.ts
@@ -1,6 +1,13 @@
-
+// Client
 export { default as Home } from "@/pages/client/Home";
 export {default as EditProfile} from "@/pages/client/EditProfile";
-export {default as EmbeddedSystem} from "@/pages/client/EmbeddedSystem"
 
+
+// Admin
+export {default as AdminHome} from "@/pages/admin/AdminHome"
+
+
+
+// Special pages
 export {default as NotFound} from "@/pages/NotFound"
+export {default as EmbeddedSystem} from "@/pages/client/EmbeddedSystem"
